fix tests

+++ src/Tar.php
@@ -189,7 +189,7 @@ class Tar implements ExtractableInterface
 
             $position += 512;
             $contents = substr($data, $position, octdec($info['size']));
-            $position += ceil(octdec($info['size']) / 512) * 512;
+            $position += (int) ceil(octdec($info['size']) / 512) * 512;
 
             if ($info['filename']) {
                 $file = [
